+++ R/BEMM.1PLG.R
@@ -3,9 +3,9 @@ BEMM.1PLG=function(data, 					#A matrix of respoonse [n.examinees * n.items]
                  PriorGamma=c(-1.39,0.25),  #The normal prior for gamma parameters with default mean -1.39 and variance 0.25
                  InitialBeta=NA,			#Initial values for beta parameters, default is NA 
                  InitialGamma=NA,			#Initial values for gamma parameters, default is NA 
-                 Tol=0.001,				    #The tolerate threshold for convergnece, default is 0.001
+                 Tol=0.0001,				#The tolerate threshold for convergnece, default is 0.0001
                  max.ECycle=2000L,		    #The max of Estem interation, default is 2000L
-                 max.MCycle=30L,		    #The max of Mstep interation, default is 30L
+                 max.MCycle=100L,		    #The max of Mstep interation, default is 100L
                  n.decimal=3L,              #The decimal length of outputs parameters, default is 3L
                  n.Quadpts =31L,		    #The number of quadratures, default is 31L
                  Theta.lim=c(-6,6),         #The range the Theta, default is [-6,6]
@@ -131,7 +131,7 @@ BEMM.1PLG=function(data, 					#A matrix of respoonse [n.examinees * n.items]
   
   
   #Compute model fit information
-  np=J*3		#Obtain the number of estimated parameters
+  np=J*2		#Obtain the number of estimated parameters
   N2loglike=-2*LogL			#-2Log-likelihood
   AIC=2*np+N2loglike		#AIC
   BIC=N2loglike+log(I)*np	#BIC
@@ -204,7 +204,7 @@ BEMM.1PLG=function(data, 					#A matrix of respoonse [n.examinees * n.items]
   }else{
     message('PROCEDURE TERMINATED WITH ISSUES')
   }
-  message('IRTEMM version: 1.0.2') 
+  message('IRTEMM version: 1.0.3') 
   message('Item Parameter Calibration for the 1PLGM.','\n')
   message('Quadrature: ', n.Quadpts, ' nodes from ', Theta.lim[1], ' to ', Theta.lim[2], ' were used to approximate Gaussian distribution.') 
   message('Method for Items: Ability-based Bayesian Expectation-Maximization-Maximization (BEMM) Algorithm.')