+++ R/BEMM.4PL.R
@@ -1,4 +1,4 @@
-BE3M.4PL=function(data, 				#A matrix of respoonse [n.examinees * n.items]
+BEMM.4PL=function(data, 				#A matrix of respoonse [n.examinees * n.items]
                  PriorA=c(0,0.25), 	    #The log normal prior for a parameters with default mean 0 and variance 0.25
                  PriorB=c(0,4),         #The normal prior for b parameters with default mean 0 and variance 4
                  PriorC=c(4,16),        #The beta prior for c parameters with default hyper-parameter 4 and 16
@@ -7,9 +7,9 @@ BE3M.4PL=function(data, 				#A matrix of respoonse [n.examinees * n.items]
                  InitialB=NA,			#Initial values for b parameters, default is NA 
                  InitialC=NA,			#Initial values for c parameters, default is NA 
                  InitialS=NA,			#Initial values for s parameters, default is NA 
-                 Tol=0.001,				#The tolerate threshold for convergnece, default is 0.001
+                 Tol=0.0001,			#The tolerate threshold for convergnece, default is 0.0001
                  max.ECycle=2000L,		#The max of Estem interation, default is 2000L
-                 max.MCycle=30L,		#The max of Mstep interation, default is 30L
+                 max.MCycle=100L,		#The max of Mstep interation, default is 100L
                  n.decimal=3L,          #The decimal length of outputs parameters, default is 3L
                  n.Quadpts =31L,		#The number of quadratures, default is 31L
                  Theta.lim=c(-6,6),     #The range the Theta, default is [-6,6]
@@ -93,9 +93,9 @@ BE3M.4PL=function(data, 				#A matrix of respoonse [n.examinees * n.items]
   n.ECycle=0L					#The first E-step iteration
   StopNormal=1L					#Whether program terminate normally
   
-  #Call BE3M estimation program from C
+  #Call BEMM estimation program from C
   
-  res=.C("BE3M4PL", data.simple=data.simple, CountNum=CountNum, n.class=n.class, J=J,  LH=LH, 
+  res=.C("BEMM4PL", data.simple=data.simple, CountNum=CountNum, n.class=n.class, J=J,  LH=LH, 
          IA=IA, IB=IB, IC=IC, IS=IS, IAB=IAB, TA=TA, TB=TB, TC=TC, TS=TS,
          deltahat.A=deltahat.A, deltahat.B=deltahat.B, deltahat.C=deltahat.C, deltahat.S=deltahat.S,
          est.A=Par.est0$A, est.B=Par.est0$B, est.C=Par.est0$C, est.S=Par.est0$S, Tol=Tol, cr=cr, StopNormal=StopNormal,
@@ -151,7 +151,7 @@ BE3M.4PL=function(data, 				#A matrix of respoonse [n.examinees * n.items]
   
   
   #Compute model fit information
-  np=J*3		#Obtain the number of estimated parameters
+  np=J*4		#Obtain the number of estimated parameters
   N2loglike=-2*LogL			#-2Log-likelihood
   AIC=2*np+N2loglike		#AIC
   BIC=N2loglike+log(I)*np	#BIC
@@ -225,10 +225,10 @@ BE3M.4PL=function(data, 				#A matrix of respoonse [n.examinees * n.items]
   }else{
     message('PROCEDURE TERMINATED WITH ISSUES')
   }
-  message('IRTEMM version: 1.0.2') 
+  message('IRTEMM version: 1.0.3') 
   message('Item Parameter Calibration for the 4PLM.','\n')
   message('Quadrature: ', n.Quadpts, ' nodes from ', Theta.lim[1], ' to ', Theta.lim[2], ' were used to approximate Gaussian distribution.') 
-  message('Method for Items: Ability-based Bayesian Expectation-Maximization-Maximization-Maximization (BE3M) Algorithm.')
+  message('Method for Items: Ability-based Bayesian Expectation-Maximization-Maximization (BEMM) Algorithm.')
   message('Method for Item SEs: Updated Supplemented EM.')
   message('Method for Theta: Expected A Posteriori (EAP).')
   if (StopNormal==1){