+++ src/BEMM3PL.c
@@ -159,15 +159,15 @@ void BEMM3PL(double *data, int *CountNum, int *n_class, int *n_item, double *LH,
 					lbb_core += wsf;          			//lbb: (f*ws)
 					lab_core += wsf * x_bt;   			//lab: wsf*(x-bt)
 				}
-				la1 = D * la1_core;
+				la1 = Da * la1_core;
 				lb1 = -Da * lb1_core;
-				laa = -D *D * laa_core;
+				laa = -D2a2 * laa_core;
 				lbb = -D2a2 * lbb_core;
-				lab = D* Da * lab_core;
+				lab = D2a2 * lab_core;
 				// Maximize a and b
 				if (PriorA[j]!=-9 && PriorA[j + J]!=-9) {
-					la1 += - 1/(at0) * ((log(at0)-PriorA[j]+PriorA[j + J])/PriorA[j + J]);
-					laa += - (1/(at0 * at0)) * (( 1 - log(at0) + PriorA[j] - PriorA[j + J]) / PriorA[j + J]);
+					la1 += -((log(at0)-PriorA[j])/PriorA[j + J]);
+					laa += -1/PriorA[j + J];
 				}
 				if (PriorB[j]!=-9 && PriorB[j + J]!=-9) {
 					lb1 += -((bt0-PriorB[j])/PriorB[j + J]);
@@ -178,19 +178,19 @@ void BEMM3PL(double *data, int *CountNum, int *n_class, int *n_item, double *LH,
 				Ibb = - laa / Weight;
 				Iab = lab / Weight;
 				Icc = - 1 / lcc;
-				at1 = at0 + (Iaa * la1 + Iab * lb1);
+				at1 = log(at0) + (Iaa * la1 + Iab * lb1);
 				bt1 = bt0 + (Ibb * lb1 + Iab * la1);
 				la1_core = 0;
 				lb1_core = 0;
 				laa_core = 0;
 				lbb_core = 0;
 				lab_core = 0;
-				if (sqrt(pow(at1-at0, 2) + pow(bt1-bt0, 2)) < 0.01){
+				if (sqrt(pow(at1-log(at0), 2) + pow(bt1-bt0, 2)) < 0.01){
 					M_exit = 0;
-					at0 = at1;
+					at0 = exp(at1);
 					bt0 = bt1;
 				} else{
-					at0 = at1;
+					at0 = exp(at1);
 					bt0 = bt1;
 					n_MCycle += 1;
 				}
@@ -303,12 +303,23 @@ void BEMM3PL(double *data, int *CountNum, int *n_class, int *n_item, double *LH,
 	double delta[5]={0};
 	double delta0[5]={0};
 	double delta1[5]={0};
+	int cr_SEM=n_ECycle[0];
+	double cr_SEM0=1;
 	double cr_SEM1=1;
 	double cr_SEM2=1;
 	double cr_SEM3=1;
-	if (n_ECycle[0]>=10){
-		start_SEM=floor(n_ECycle[0] * 0.2);
-		end_SEM=floor(n_ECycle[0] * 0.8);
+	double deltatemp;
+	
+	for (i = 0; i<=cr_SEM; i++){
+		deltatemp=exp(-(LH[i+1]-LH[i]));
+		if (deltatemp>=0.9 && deltatemp<=0.999){
+			if (cr_SEM0==0){
+				end_SEM=i;
+			}else{
+				start_SEM=i;
+				cr_SEM0=0;
+			}
+		}
 	}
 	for (jj = 0; jj < J; jj++) {
 		z=start_SEM;
@@ -319,13 +330,13 @@ void BEMM3PL(double *data, int *CountNum, int *n_class, int *n_item, double *LH,
 		while (SEM_exit && z<=end_SEM){
 			mm = jj + J * z;
 			for (ParClass = 1; ParClass <= 3; ParClass++){
-				if (ParClass==1 && z>=2 && cr_SEM1<0.01){
+				if (ParClass==1 && z>=2 && cr_SEM1<0.0001){
 					continue;
 				}
-				if (ParClass==2 && z>=2 && cr_SEM2<0.01){
+				if (ParClass==2 && z>=2 && cr_SEM2<0.0001){
 					continue;
 				}
-				if (ParClass==3 && z>=2 && cr_SEM3<0.01){
+				if (ParClass==3 && z>=2 && cr_SEM3<0.0001){
 					continue;
 				}
 				for (j = 0; j < J; j++){
@@ -429,38 +440,38 @@ void BEMM3PL(double *data, int *CountNum, int *n_class, int *n_item, double *LH,
 							lbb_core += wsf;          			//lbb: (f*ws)
 							lab_core += wsf * x_bt;   			//lab: wsf*(x-bt)
 						}
-						la1 = D * la1_core;
+						la1 = Da * la1_core;
 						lb1 = -Da * lb1_core;
-						laa = -D *D * laa_core;
+						laa = -D2a2 * laa_core;
 						lbb = -D2a2 * lbb_core;
-						lab = D* Da * lab_core;
+						lab = D2a2 * lab_core;
 						// Maximize a and b
-						if (PriorA[jj]!=-9 && PriorA[jj + J]!=-9) {
-							la1 += - 1/(at0) * ((log(at0)-PriorA[jj]+PriorA[jj + J])/PriorA[jj + J]);
-							laa += - (1/(at0 * at0)) * (( 1 - log(at0) + PriorA[jj] - PriorA[jj + J]) / PriorA[jj + J]);
+						if (PriorA[j]!=-9 && PriorA[j + J]!=-9) {
+							la1 += -((log(at0)-PriorA[j])/PriorA[j + J]);
+							laa += -1/PriorA[j + J];
 						}
-						if (PriorB[jj]!=-9 && PriorB[jj + J]!=-9) {
-							lb1 += -((bt0-PriorB[jj])/PriorB[jj + J]);
-							lbb += -1/PriorB[jj + J];
+						if (PriorB[j]!=-9 && PriorB[j + J]!=-9) {
+							lb1 += -((bt0-PriorB[j])/PriorB[j + J]);
+							lbb += -1/PriorB[j + J];
 						}
 						Weight = laa * lbb - lab * lab;
 						Iaa = - lbb / Weight;
 						Ibb = - laa / Weight;
 						Iab = lab / Weight;
 						Icc = - 1 / lcc;
-						at1 = at0 + (Iaa * la1 + Iab * lb1);
+						at1 = log(at0) + (Iaa * la1 + Iab * lb1);
 						bt1 = bt0 + (Ibb * lb1 + Iab * la1);
 						la1_core = 0;
 						lb1_core = 0;
 						laa_core = 0;
 						lbb_core = 0;
 						lab_core = 0;
-						if (sqrt(pow(at1-at0, 2) + pow(bt1-bt0, 2)) < 0.01){
+						if (sqrt(pow(at1-log(at0), 2) + pow(bt1-bt0, 2)) < 0.01){
 							M_exit = 0;
-							at0 = at1;
+							at0 = exp(at1);
 							bt0 = bt1;
 						} else{
-							at0 = at1;
+							at0 = exp(at1);
 							bt0 = bt1;
 							n_MCycle += 1;
 						}
@@ -481,22 +492,22 @@ void BEMM3PL(double *data, int *CountNum, int *n_class, int *n_item, double *LH,
 				}
 				switch (ParClass){
 					case 1:
-						delta1[0]=(deltahat_A[jj]-A[jj])/(TA[mm]-A[jj]);
-						delta1[1]=(deltahat_B[jj]-B[jj])/(TA[mm]-A[jj]);
+						delta1[0]=(log(deltahat_A[jj])-log(A[jj]))/(log(TA[mm])-log(A[jj])+0.0001);
+						delta1[1]=(deltahat_B[jj]-B[jj])/(log(TA[mm])-log(A[jj])+0.0001);
 						break;
 					case 2:
-						delta1[2]=(deltahat_A[jj]-A[jj])/(TB[mm]-B[jj]);
-						delta1[3]=(deltahat_B[jj]-B[jj])/(TB[mm]-B[jj]);
+						delta1[2]=(log(deltahat_A[jj])-log(A[jj]))/(TB[mm]-B[jj]+0.0001);
+						delta1[3]=(deltahat_B[jj]-B[jj])/(TB[mm]-B[jj]+0.0001);
 						break;
 					case 3:
-						delta1[4]=(deltahat_C[jj]-C[jj])/(TC[mm]-C[jj]);   
+						delta1[4]=(deltahat_C[jj]-C[jj])/(TC[mm]-C[jj]+0.0001);   
 						break;						
 				}
 			}
 			cr_SEM1=sqrt(pow(delta1[0]-delta0[0],2)+pow(delta1[1]-delta0[1],2));
 			cr_SEM2=sqrt(pow(delta1[2]-delta0[2],2)+pow(delta1[3]-delta0[3],2));
 			cr_SEM3=fabs(delta1[4]-delta0[4]);
-			if (cr_SEM1<0.01 && cr_SEM2<0.01 && cr_SEM3<0.01 && z>=2){
+			if (cr_SEM1<0.0001 && cr_SEM2<0.0001 && cr_SEM3<0.0001 && z>=2){
 				SEM_exit=0;
 			}else{
 				z=z+1;
@@ -518,14 +529,17 @@ void BEMM3PL(double *data, int *CountNum, int *n_class, int *n_item, double *LH,
 		delta1[2] = -delta[2] / Weight;
 		delta1[3] =  delta[0] / Weight;
 		delta1[4] =  1 / delta[4];
-		if (isnormal(delta1[0])==0){delta1[0] = 1;}
-		if (isnormal(delta1[1])==0){delta1[1] = 0;}
-		if (isnormal(delta1[2])==0){delta1[2] = 0;}
-		if (isnormal(delta1[3])==0){delta1[3] = 1;}
-		if (isnormal(delta1[4])==0){delta1[4] = 1;}
-		SEA[jj]= sqrt(fabs(IA[jj] * delta1[0] + IAB[jj] * delta1[2]));
-        SEB[jj]= sqrt(fabs(IB[jj] * delta1[3] + IAB[jj] * delta1[1]));
-        SEC[jj]= sqrt(fabs(IC[jj] * delta1[4]));
+		if (isnormal(delta1[0])==0 || delta1[0]<=0){delta1[0] = 1;}
+		if (isnormal(delta1[1])==0 || delta1[1]<=0){delta1[1] = 0;}
+		if (isnormal(delta1[2])==0 || delta1[2]<=0){delta1[2] = 0;}
+		if (isnormal(delta1[3])==0 || delta1[3]<=0){delta1[3] = 1;}
+		if (isnormal(delta1[4])==0 || delta1[4]<=0){delta1[4] = 1;}
+		SEA[jj]= sqrt(A[jj] * A[jj] * IA[jj] * delta1[0] + IAB[jj] * delta1[2]);
+        SEB[jj]= sqrt(IB[jj] * delta1[3] + IAB[jj] * delta1[1]);
+        SEC[jj]= sqrt(IC[jj] * delta1[4]);
+		if (SEA[jj]>1){SEA[jj]= sqrt(A[jj] * A[jj] * IA[jj]);}
+		if (SEB[jj]>1){SEB[jj]= sqrt(IB[jj]);}
+		if (SEC[jj]>1){SEC[jj]= sqrt(IC[jj]);}
 	}
 	n_ECycle[0] = n_ECycle[0] + 1;
 }